+++ src/main/java/woowacourse/shoppingcart/application/CartService.java
@@ -7,8 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import woowacourse.shoppingcart.dao.CartItemDao;
 import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.dao.ProductDao;
-import woowacourse.shoppingcart.domain.CartItem;
-import woowacourse.shoppingcart.domain.Quantity;
+import woowacourse.shoppingcart.domain.cart.CartItem;
 import woowacourse.shoppingcart.dto.CartItemResponse;
 import woowacourse.shoppingcart.dto.ProductResponse;
 import woowacourse.shoppingcart.exception.CartItemNotFoundException;
