+++ src/main/java/woowacourse/shoppingcart/domain/cart/Quantity.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.domain;
+package woowacourse.shoppingcart.domain.cart;
 
 import woowacourse.shoppingcart.exception.InvalidFormException;
