+++ src/test/java/woowacourse/shoppingcart/dao/CartItemItemDaoTest.java
@@ -12,8 +12,8 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.jdbc.Sql;
-import woowacourse.shoppingcart.domain.CartItem;
-import woowacourse.shoppingcart.domain.Quantity;
+import woowacourse.shoppingcart.domain.cart.CartItem;
+import woowacourse.shoppingcart.domain.cart.Quantity;
 import woowacourse.shoppingcart.domain.product.Product;
 import woowacourse.shoppingcart.dto.ProductRequest;
